moved all executable scripts to scripts directory
