+++ readnext/modeling/__init__.py
@@ -1,8 +1,20 @@
+from readnext.modeling.citation_models_features import (
+    add_feature_rank_cols,
+    set_missing_publication_dates_to_max_rank,
+)
 from readnext.modeling.document_info import (
     DocumentInfo,
     DocumentScore,
     DocumentsInfo,
     documents_info_from_df,
+)
+from readnext.modeling.embedder import (
+    BERTEmbedder,
+    BM25Embedder,
+    FastTextEmbedder,
+    TFIDFEmbedder,
+    Word2VecEmbedder,
+    embeddings_mapping_to_frame,
 )
 from readnext.modeling.model_data import (
     CitationModelData,
@@ -14,16 +26,40 @@ from readnext.modeling.model_data_constructor import (
     LanguageModelDataConstructor,
     ModelDataConstructor,
 )
+from readnext.modeling.tokenizer import (
+    BERTTokenizer,
+    SpacyTokenizer,
+    StringMapping,
+    TokenIds,
+    Tokens,
+    TokensIdMapping,
+    TokensMapping,
+)
 
 __all__ = [
+    "add_feature_rank_cols",
+    "set_missing_publication_dates_to_max_rank",
     "DocumentInfo",
     "DocumentScore",
     "DocumentsInfo",
     "documents_info_from_df",
+    "BERTEmbedder",
+    "BM25Embedder",
+    "FastTextEmbedder",
+    "TFIDFEmbedder",
+    "Word2VecEmbedder",
+    "embeddings_mapping_to_frame",
     "CitationModelData",
     "LanguageModelData",
     "ModelData",
     "CitationModelDataConstructor",
     "LanguageModelDataConstructor",
     "ModelDataConstructor",
+    "BERTTokenizer",
+    "SpacyTokenizer",
+    "StringMapping",
+    "TokenIds",
+    "Tokens",
+    "TokensIdMapping",
+    "TokensMapping",
 ]
