--- readnext/modeling/citation_models/__init__.py
@@ -1,6 +1,0 @@
-from readnext.modeling.citation_models.features import (
-    add_feature_rank_cols,
-    set_missing_publication_dates_to_max_rank,
-)
-
-__all__ = ["add_feature_rank_cols", "set_missing_publication_dates_to_max_rank"]
