+++ readnext/scripts/data/s6_merge_authors.py
