+++ tests/test_data/test_results_tokenized_abstracts.py
@@ -1,7 +1,7 @@
 import pytest
 
 from readnext.config import ResultsPaths
-from readnext.modeling.language_models import TokensIdMapping, TokensMapping
+from readnext.modeling import TokensIdMapping, TokensMapping
 from readnext.utils import load_object_from_pickle, slice_mapping
