+++ tests/test_models/test_citation_models_features.py
@@ -2,7 +2,7 @@ import pandas as pd
 import pytest
 from pandas.testing import assert_frame_equal
 
-from readnext.modeling.citation_models import (
+from readnext.modeling import (
     add_feature_rank_cols,
     set_missing_publication_dates_to_max_rank,
 )
