+++ tests/test_tokenizers/test_bert_tokenizer.py
@@ -3,7 +3,7 @@ from transformers import BertTokenizerFast
 
 from readnext.config import ModelVersions
 from readnext.modeling import DocumentsInfo
-from readnext.modeling.language_models import BERTTokenizer
+from readnext.modeling import BERTTokenizer
 
 
 @pytest.fixture(scope="module")
